refactor: add inheritance-based hierarchy to chart views
Refactors chart view system from single monolithic file to modular
inheritance-based structure. Views now inherit from a base config and
only override what's different.

Structure:
```
app/lib/chart/chartViews/
├── index.ts          # Main exports + mergeWithBase()
├── types.ts          # ChartContext, ReferenceLineConfig, etc.
├── helpers.ts        # Shared utility functions
├── base.ts           # BASE_VIEW with default implementations
├── mortality.ts      # Inherits all from base (no overrides)
├── excess.ts         # Overrides titles + adds baseline line
└── zscore.ts         # Overrides most for statistical view
```

Benefits:
- DRY: Shared logic in base, views only define differences
- Consistency: All views inherit same base behavior
- Easy to extend: New views just specify overrides
- Maintainable: Each view is isolated in its own file
- Mirrors state library pattern (views.ts, viewTypes.ts, etc.)

Example - Mortality view:
```typescript
export const MORTALITY_VIEW: ChartViewConfig = {
  // Inherits everything from BASE_VIEW
  // Zero overrides needed!
}
```

Example - Excess view:
```typescript
export const EXCESS_VIEW: ChartViewConfig = {
  getTitleParts: (ctx) => [...], // Override: Add "Excess" prefix
  yAxisLabel: 'Excess Deaths',    // Override: Excess labels
  referenceLines: () => [...]     // Add: Baseline at y=0
  // Inherits getSubtitle, etc. from base
}
```

No breaking changes - all tests pass.
