+++ app/lib/chart/chartViews/mortality.ts
@@ -1,0 +1,17 @@
+/**
+ * Mortality View Configuration
+ *
+ * Standard mortality data visualization
+ * Uses base config with no overrides (default behavior)
+ */
+
+import type { ChartViewConfig } from './types'
+
+/**
+ * Mortality view - inherits all behavior from base
+ * This is the default view with standard mortality visualization
+ */
+export const MORTALITY_VIEW: ChartViewConfig = {
+  // Inherits everything from BASE_VIEW
+  // No overrides needed - base provides perfect defaults
+}
